requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cleanup==8.0.0
django-jalali-date==1.0.2
django-render-partial==0.4
idna==3.4
jdatetime==4.1.1
Pillow==10.0.0
psycopg2==2.9.6
requests==2.31.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
